@@ -1,18 +1,24 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
 
     defaultConfig {
         applicationId "at.huber.sampleDownload"
-        minSdkVersion 14
-        //noinspection ExpiredTargetSdkVersion
-        targetSdkVersion 22
+        minSdkVersion 21
+        targetSdkVersion 29
         versionCode 2
         versionName "2.1"
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 }
 
 dependencies {
     implementation project(':youtubeExtractor')
+
+    implementation 'androidx.appcompat:appcompat:1.2.0'
 }
